% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{chooseNonHierarchical}
\alias{chooseNonHierarchical}
\title{CHOOSE NON HIERARCHICAL}
\usage{
chooseNonHierarchical(dataset_name, number_folds, similarity, folderResults)
}
\arguments{
\item{dataset_name}{Name of the dataset to be analyzed.}

\item{number_folds}{Number of folds for cross-validation.}

\item{similarity}{Method of similarity used in the analysis (e.g., "knn", "threshold").}

\item{folderResults}{Path to the folder where results will be saved.}
}
\value{
No value is returned by the function, but CSV files containing the selected partitions for each
fold are saved in the directory specified in \code{folderResults}.
}
\description{
This function selects non-hierarchical partitions for each dataset, based on different types of
sparsification and partitioning methods. It is applied to multiple datasets, divided into several
folds for cross-validation.
}
\details{
The function performs the selection of non-hierarchical partitions, both for k-Nearest Neighbors (KNN)
and Thresholds, considering different conditions such as modularity values. The process includes creating
directories, reading and filtering data files, and generating CSV files with the selected partitions for
each fold.

The function follows these steps:
\enumerate{
\item Creates specific directories to store the results.
\item Reads CSV files containing information about modularity and partitions.
\item Filters to select non-hierarchical and hybrid partitions.
\item Selects the partitions with the highest modularity and chooses an appropriate method.
\item Saves the results for each fold in CSV files, including the selected partitions.
}
}
\examples{
\dontrun{
chooseNonHierarchical("dataset1", 5, "knn", "/path/to/results")
}

}
\seealso{
\code{\link{chooseHierarchical}} for the function that selects hierarchical partitions.
}
