% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{communities.detections.names}
\alias{communities.detections.names}
\title{Get Community Detection Method Names}
\usage{
communities.detections.names()
}
\value{
A character vector with the names of community detection methods.
}
\description{
This function returns a vector containing the names of various community
detection algorithms available in the \code{igraph} package.
}
\examples{
methods <- communities.detections.names()
print(methods)

}
\seealso{
\code{igraph::cluster_spinglass}, \code{igraph::cluster_edge_betweenness},
\code{igraph::cluster_fast_greedy}, \code{igraph::cluster_infomap},
\code{igraph::cluster_label_prop}, \code{igraph::cluster_leading_eigen},
\code{igraph::cluster_louvain}, \code{igraph::cluster_optimal},
\code{igraph::cluster_walktrap}
}
