% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.knn}
\alias{compute.knn}
\title{Compute k-Nearest Neighbors (kNN) Results}
\usage{
\method{compute}{knn}(parameters)
}
\arguments{
\item{parameters}{A list containing necessary parameters including dataset name, folder paths, similarity values, and control flags.}
}
\value{
A list containing runtime and summary results for different kNN values.
}
\description{
This function processes a set of kNN parameters and executes multiple community detection algorithms on graph data.
It stores runtime and summary results, saving them as CSV files if specified.
}
\examples{
parameters <- list(knn.values = list(values = c(5, 10, 15)),
                    FolderSplit = "./data",
                    FolderSplitDF = "./dataframes",
                    Dataset.Name = "example",
                    Dataset.Info = list(Labels = 3),
                    fold = 1,
                    similarity = "cosine",
                    Save.Csv.Files = 1)
result <- compute.knn(parameters)

}
