% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.tr}
\alias{compute.tr}
\title{Compute Training Results}
\usage{
\method{compute}{tr}(parameters)
}
\arguments{
\item{parameters}{A list containing necessary parameters including dataset name, folder paths, similarity values, and control flags.}
}
\value{
A list containing runtime and summary results for different training values.
}
\description{
This function processes a set of training parameters and executes multiple community detection algorithms on graph data.
It stores the runtime and summary of results, saving them as CSV files if specified.
}
\examples{
parameters <- list(tr.values = list(values = c(0.1, 0.2, 0.3)),
                    FolderSplit = "./data",
                    FolderSplitDF = "./dataframes",
                    Dataset.Name = "example",
                    Dataset.Info = list(Labels = 3),
                    fold = 1,
                    similarity = "cosine",
                    Save.Csv.Files = 1)
result <- compute.tr(parameters)

}
