% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{execute.communities}
\alias{execute.communities}
\title{Execute community detection algorithms on multiple dataset folds}
\usage{
execute.communities(parameters)
}
\arguments{
\item{parameters}{A list containing dataset and configuration details.}
}
\value{
A list containing execution times, summaries, and results of multiple community detection methods.
}
\description{
This function iterates over multiple dataset folds and applies different community detection methods.
It stores execution times and summaries for each fold.
}
\examples{

parameters <- list(
  Number.Folds = 5,
  similarity = "jaccard",
  Dataset.Name = "my_dataset",
  Dataset.Info = list(Labels = 3),
  Folders = list(folderCommunities = "~/Communities", folderLabelGraphs = "~/Graphs")
)

results <- execute_communities(parameters)

}
