% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_file}
\alias{extract_file}
\title{Function to check the compression type and extract the file}
\usage{
extract_file(file_path, dest_dir)
}
\arguments{
\item{file_path}{The path of the file to be extracted.}

\item{dest_dir}{The directory where the file will be extracted to.}
}
\value{
None. Stops execution if an error occurs during extraction.
}
\description{
This function checks the compression type of a file and extracts it accordingly.
}
