% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\docType{data}
\name{FolderRoot}
\alias{FolderRoot}
\title{Generate Partitions Communities}
\format{
An object of class \code{character} of length 1.
}
\usage{
FolderRoot
}
\description{
This script sets up the environment for generating partition communities by
installing and loading necessary R packages. It ensures that required packages
are installed before use.
}
\details{
The script defines the root folder for storing scripts and data,
installs missing packages from a predefined list, and loads the required
libraries. The packages used include \code{igraph}, \code{dplyr}, \code{tidyr}, \code{stringr},
\code{parallel}, and \code{doParallel}, which are essential for graph-based community
detection and parallel computing.
}
\note{
Any errors, please contact: elainececiliagatto@gmail.com
}
\examples{
# Running the script ensures required packages are installed and loaded

}
\seealso{
\code{install.packages}, \code{library}, \code{require}
}
\author{
PhD Elaine Cecilia Gatto
Federal University of Lavras (UFLA), Applied Computer Department (DAC)

PhD Ricardo Cerri
State University os São Paulo

PhD Mauri Ferrandin
Federal University of Santa Catarina

PhD Alan Demetrius
Federal University of Sao Carlos (UFSCar), Computer Department (DC)
}
\keyword{datasets}
