% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.parameters}
\alias{get.parameters}
\title{Get Parameters}
\usage{
get.parameters()
}
\value{
The list of stored parameters.
}
\description{
Retrieves the currently stored parameters.
}
\note{
This function currently does not store parameters internally, so it will not work as expected.
}
\examples{
get.parameters()
}
