% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{install_if_missing}
\alias{install_if_missing}
\title{Install missing packages and load them}
\usage{
install_if_missing(pkg)
}
\arguments{
\item{pkg}{Character string with the package name.}
}
\value{
Loads the package after installation.
}
\description{
Install missing packages and load them
}
\examples{
install_if_missing("igraph")
}
