% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{juntaDFs}
\alias{juntaDFs}
\title{JOIN DATA FRAMES}
\usage{
juntaDFs(dataset_name, number_folds, similarity, folderResults)
}
\arguments{
\item{dataset_name}{Name of the dataset to be processed.}

\item{number_folds}{Number of folds to process.}

\item{similarity}{A string indicating the similarity method used in the dataset.}

\item{folderResults}{Path to the folder where results will be stored.}
}
\value{
This function does not return any value. It writes two CSV files:
"All-Conectado-KNN.csv" and "All-Conectado-TR.csv" containing the combined data.
}
\description{
This function consolidates multiple fold-specific CSV files into two large summary data frames.
The function reads KNN (k-Nearest Neighbors) and Threshold (TR) CSV files for each fold and appends
them into one combined data frame for each type. The resulting combined data frames are saved as CSV
files in the specified folder.
}
\details{
The function iterates over the specified number of folds (usually from 1 to 10), reading the KNN and
Threshold files generated for each fold, and appending them to the \code{resumeKNN} and \code{resumeTR} data frames,
respectively. After all folds are processed, the combined data frames are written to CSV files.

The function performs the following steps:
\enumerate{
\item Creates necessary directories for the dataset.
\item Loops through each fold, reading the KNN and Threshold CSV files.
\item Appends the data from each fold to the appropriate data frames (\code{resumeKNN} and \code{resumeTR}).
\item After processing all folds, writes the combined data frames to CSV files.
}
}
\examples{
\dontrun{
juntaDFs("dataset1", 10, "cosine", "/path/to/results")
}

}
\seealso{
\code{\link{createDirs2}} for the function that creates the necessary directories.
}
