% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_datasets}
\alias{load_datasets}
\title{Load Datasets Information}
\usage{
load_datasets()
}
\value{
A dataframe containing dataset metadata.
}
\description{
Reads the dataset metadata from a CSV file and returns it as a dataframe.
}
\examples{
datasets <- load_datasets()
head(datasets)
}
