% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{moveFilesPartitions}
\alias{moveFilesPartitions}
\title{MOVE FILES PARTITIONS}
\usage{
moveFilesPartitions(dataset_name, number_folds, folderResults)
}
\arguments{
\item{dataset_name}{Name of the dataset to be processed.}

\item{number_folds}{Number of folds used in cross-validation.}

\item{folderResults}{Path to the folder where the results will be stored.}
}
\value{
This function does not return any value. It performs file copying operations to organize the
generated partition files into the correct directories.
}
\description{
This function is designed to move partition files generated during the process of partitioning datasets
for multiple folds in cross-validation. The function copies the relevant partition files for both
k-Nearest Neighbors (KNN) and Threshold methods, as well as the associated hybrid partitions, from
their source directories to destination directories for each fold.
}
\details{
The function handles both KNN and Threshold partitioning methods. It first checks for the existence of
specific CSV files related to KNN and Thresholds and then copies them into the appropriate subdirectories
for each fold. Additionally, it creates any necessary directories if they do not already exist.

The function operates as follows:
\enumerate{
\item Creates the necessary directories for each fold.
\item Copies partition files for both KNN and Threshold methods, including hierarchical and non-hierarchical partitions.
\item Checks the existence of certain dataframes before proceeding with the file copying.
\item Ensures that each partition type is copied to the correct fold-specific subdirectory.
}
}
\examples{
\dontrun{
moveFilesPartitions("dataset1", 5, "/path/to/results")
}

}
\seealso{
\code{\link{chooseNonHierarchical}} for the function that selects non-hierarchical partitions.
}
