% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot Graphs}
\usage{
plotGraph(community, graph, title)
}
\arguments{
\item{community}{An object containing the community detection result.}

\item{graph}{An igraph object representing the graph structure.}

\item{title}{A string specifying the title of the plot.}
}
\value{
A plot displaying the graph with community structure.
}
\description{
This function plots a graph with a given community detection result.
It visualizes the community structure with customized aesthetics.
}
\examples{
plotGraph(community, graph, "Community Structure")
}
