% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_dataset}
\alias{process_dataset}
\title{Process Dataset}
\usage{
process_dataset(datasets, parameters)
}
\arguments{
\item{datasets}{A dataframe containing dataset information.}

\item{parameters}{A list containing configuration settings, including paths.}
}
\value{
None. The function stops execution if the dataset file is missing.
}
\description{
This function checks if the dataset file exists, copies it to the appropriate folder,
extracts its contents, and organizes necessary files for processing.
}
\examples{
process_dataset(datasets, parameters)
}
