% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_label_graphs}
\alias{process_label_graphs}
\title{Process Similarities Matrices}
\usage{
process_label_graphs(datasets, parameters)
}
\arguments{
\item{datasets}{A dataframe containing similarity information.}

\item{parameters}{A list containing configuration settings, including paths.}
}
\value{
None. Stops execution if the similarity file is missing.
}
\description{
Checks if the similarity file exists, and if so, proceeds with copying and extracting it.
}
