% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_community_method}
\alias{run_community_method}
\title{Run a Community Detection Method}
\usage{
run_community_method(method_function, parameters, graph, title, fold)
}
\arguments{
\item{method_function}{A function that performs a community detection algorithm.
It should accept \code{parameters}, \code{graph}, \code{title}, \code{fold}, and \code{Save} as arguments.}

\item{parameters}{A list of parameters required for the community detection method.
It should contain at least the \code{FolderSplit} path where results will be saved.}

\item{graph}{An \code{igraph} object representing the network structure on which
the community detection method will be applied.}

\item{fold}{An integer indicating the current fold number in a cross-validation
or iterative process.}
}
\value{
A numeric vector of length 3 containing execution time statistics:
- \code{user}: Time spent in user mode.
- \code{system}: Time spent in system mode.
- \code{elapsed}: Total elapsed time.
}
\description{
This function executes a specified community detection algorithm on a given graph
and measures its execution time.
}
\examples{
\dontrun{
parameters <- list(FolderSplit = "results/fold1")
graph <- igraph::make_ring(10)
result <- run_community_method(executeLouvain, parameters, graph, 1)
print(result)
}

}
