% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.parameters}
\alias{set.parameters}
\title{Set Parameters}
\usage{
set.parameters(parameters)
}
\arguments{
\item{parameters}{A list containing configuration settings.}
}
\value{
The same list of parameters, unmodified.
}
\description{
Stores and returns the given parameter list.
}
\examples{
params <- list(a = 1, b = 2)
set.parameters(params)
}
