% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifyPartition}
\alias{verifyPartition}
\title{Verify Partition}
\usage{
verifyPartition(parameters, size, vertices)
}
\arguments{
\item{parameters}{A list containing relevant parameters for partition verification.}

\item{size}{An integer representing the number of detected communities.}

\item{vertices}{An integer representing the total number of vertices in the graph.}
}
\value{
A list with partition information, including the number of communities and the partition type.
}
\description{
This function categorizes the partitioning of a graph based on the number of communities.
It determines whether the partition is global, local, or hybrid.
}
\examples{
verifyPartition(parameters, size = 3, vertices = 10)
}
